+++ stripe/api_resources/financial_connections/__init__.py
@@ -11,4 +11,8 @@ from stripe.api_resources.financial_connections.account_owner import (
 from stripe.api_resources.financial_connections.account_ownership import (
     AccountOwnership,
 )
+from stripe.api_resources.financial_connections.inferred_balance import (
+    InferredBalance,
+)
 from stripe.api_resources.financial_connections.session import Session
+from stripe.api_resources.financial_connections.transaction import Transaction