+++ stripe/api_resources/financial_connections/inferred_balance.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# File generated from our OpenAPI spec
+from __future__ import absolute_import, division, print_function
+
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class InferredBalance(ListableAPIResource):
+    """
+    A historical balance for the account on a particular day. It may be sourced from a balance snapshot provided by a financial institution, or inferred using transactions data.
+    """
+
+    OBJECT_NAME = "financial_connections.account_inferred_balance"