+++ stripe/api_resources/financial_connections/transaction.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# File generated from our OpenAPI spec
+from __future__ import absolute_import, division, print_function
+
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class Transaction(ListableAPIResource):
+    """
+    A Transaction represents a real transaction that affects a Financial Connections Account balance.
+    """
+
+    OBJECT_NAME = "financial_connections.transaction"