+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 5.1.0 - 2023-02-02
+* [#920](https://github.com/stripe/stripe-python/pull/920) API Updates
+  * Add support for `resume` method on resource `Subscription`
+* [#913](https://github.com/stripe/stripe-python/pull/913) Pass params into logger.{info,debug}
+
 ## 5.1.0b7 - 2023-01-26
 * [#917](https://github.com/stripe/stripe-python/pull/917) API Updates for beta branch
   * Updated stable APIs to the latest version